Rename `pieces.py` to `piece.py`
